requirements.txt
altgraph==0.17.4
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
et-xmlfile==1.1.0
Flask==3.0.3
fonttools==4.50.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.3.2
kiwisolver==1.4.5
lxml==5.2.2
MarkupSafe==2.1.5
matplotlib==3.8.3
numpy==1.26.4
opencv-python==4.9.0.80
openpyxl==3.1.2
packaging==24.0
pandas==2.2.1
pefile==2023.2.7
pillow==10.3.0
pyinstaller==6.7.0
pyinstaller-hooks-contrib==2024.6
PyMuPDF==1.24.4
PyMuPDFb==1.24.3
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.1
pywin32-ctypes==0.2.2
requests==2.31.0
scikit-learn==1.4.1.post1
scipy==1.12.0
setuptools==70.0.0
six==1.16.0
threadpoolctl==3.4.0
tk==0.1.0
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.3
